# frontend/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cortex API",
    "description": "Evidence-first operational intelligence platform for supply chains",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/sources/upload": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Upload Source File",
        "description": "Upload a source file (CSV/XLSX). Validates, stores immutably, profiles, and creates batch/file records.",
        "operationId": "upload_source_file_api_v1_sources_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_source_file_api_v1_sources_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/batches/{batch_id}": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Get Batch Status",
        "description": "Get batch status and metadata (workspace-scoped).",
        "operationId": "get_batch_status_api_v1_sources_batches__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/files/{file_id}/profile": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Get File Profile",
        "description": "Get file profiling results (workspace-scoped).",
        "operationId": "get_file_profile_api_v1_sources_files__file_id__profile_get",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/batches/{batch_id}/claims": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List Claims",
        "description": "List evidence claims for a batch (workspace-scoped).",
        "operationId": "list_claims_api_v1_sources_batches__batch_id__claims_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClaimResponse"
                  },
                  "title": "Response List Claims Api V1 Sources Batches  Batch Id  Claims Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/batches/{batch_id}/conflicts": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List Conflicts",
        "description": "List conflicts for a batch (workspace-scoped).",
        "operationId": "list_conflicts_api_v1_sources_batches__batch_id__conflicts_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConflictResponse"
                  },
                  "title": "Response List Conflicts Api V1 Sources Batches  Batch Id  Conflicts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/conflicts/{conflict_id}/resolve": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Resolve Conflict Endpoint",
        "description": "Resolve a conflict by selecting a claim or taking an action (workspace-scoped).",
        "operationId": "resolve_conflict_endpoint_api_v1_sources_conflicts__conflict_id__resolve_post",
        "parameters": [
          {
            "name": "conflict_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conflict Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Resolve Conflict Endpoint Api V1 Sources Conflicts  Conflict Id  Resolve Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/batches/{batch_id}/compile": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Compile Batch",
        "description": "Compile a batch: extract claims, detect conflicts, compute readiness.",
        "operationId": "compile_batch_api_v1_sources_batches__batch_id__compile_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Compile Batch Api V1 Sources Batches  Batch Id  Compile Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchStatusResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "file_count": {
            "type": "integer",
            "title": "File Count"
          },
          "checksum": {
            "type": "string",
            "title": "Checksum"
          },
          "source_system_hint": {
            "type": "string",
            "title": "Source System Hint"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "workspace_id",
          "status",
          "file_count",
          "checksum",
          "source_system_hint"
        ],
        "title": "BatchStatusResponse"
      },
...
      "Body_upload_source_file_api_v1_sources_upload_post": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "source_system_hint": {
            "type": "string",
            "title": "Source System Hint",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "file"
        ],
        "title": "Body_upload_source_file_api_v1_sources_upload_post"
      },
...
      "ClaimResponse": {
        "properties": {
          "claim_id": {
            "type": "string",
            "title": "Claim Id"
          },
          "source_column": {
            "type": "string",
            "title": "Source Column"
          },
          "canonical_entity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Entity"
          },
          "canonical_field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Field"
          },
          "raw_value": {
            "type": "string",
            "title": "Raw Value"
          },
          "normalized_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Value"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "claim_state": {
            "type": "string",
            "title": "Claim State"
          }
        },
        "type": "object",
        "required": [
          "claim_id",
          "source_column",
          "canonical_entity",
          "canonical_field",
          "raw_value",
          "normalized_value",
          "confidence",
          "claim_state"
        ],
        "title": "ClaimResponse"
      },
...
      "ConflictResponse": {
        "properties": {
          "conflict_id": {
            "type": "string",
            "title": "Conflict Id"
          },
          "claim_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Claim Ids"
          },
          "canonical_entity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Entity"
          },
          "canonical_field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Field"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "blocking": {
            "type": "boolean",
            "title": "Blocking"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "conflict_id",
          "claim_ids",
          "canonical_entity",
          "canonical_field",
          "severity",
          "blocking",
          "status",
          "explanation"
        ],
        "title": "ConflictResponse"
      },
...
      "FileProfileResponse": {
        "properties": {
          "file_id": {
            "type": "string",
            "title": "File Id"
          },
          "original_name": {
            "type": "string",
            "title": "Original Name"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          },
          "column_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column Count"
          },
          "profiled": {
            "type": "boolean",
            "title": "Profiled"
          },
          "columns": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Columns"
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "original_name",
          "mime_type",
          "size",
          "row_count",
          "column_count",
          "profiled",
          "columns"
        ],
        "title": "FileProfileResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResolveRequest": {
        "properties": {
          "selected_claim_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Claim Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "approved",
              "rejected",
              "deferred",
              "merge"
            ],
            "title": "Action",
            "default": "approved"
          },
          "rationale": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Rationale"
          }
        },
        "type": "object",
        "required": [
          "rationale"
        ],
        "title": "ResolveRequest",
        "description": "Payload for conflict resolution.\n\n``rationale`` must be non-empty \u2014 humans must record a reason. ``action``\nis constrained to the known resolution verbs; anything else is a 422."
      },
...
      "UploadResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "file_id": {
            "type": "string",
            "title": "File Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "checksum": {
            "type": "string",
            "title": "Checksum"
          },
          "is_duplicate": {
            "type": "boolean",
            "title": "Is Duplicate"
          },
          "profiling_scheduled": {
            "type": "boolean",
            "title": "Profiling Scheduled"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "file_id",
          "status",
          "checksum",
          "is_duplicate",
          "profiling_scheduled"
        ],
        "title": "UploadResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}